+++ packages/state-manager/src/sagas/identity/updateCertificate/updateCertificate.saga.ts
@@ -2,12 +2,12 @@ import { SendCertificatesResponse } from '@quiet/types'
 import { PayloadAction } from '@reduxjs/toolkit'
 import { select, call, put } from 'typed-redux-saga'
 import { identitySelectors } from '../identity.selectors'
-import { CertFieldsTypes, getCertFieldValue, getReqFieldValue, loadCSR, parseCertificate } from '@quiet/identity'
+import { loadCSR, pubKeyMatch } from '@quiet/identity'
 import { identityActions } from '../identity.slice'
 import { communitiesSelectors } from '../../communities/communities.selectors'
 
 export function* updateCertificateSaga(action: PayloadAction<SendCertificatesResponse>): Generator {
-  const certificate = yield* select(identitySelectors.communityMembership)
+  const certificate = yield* select(identitySelectors.hasCertificate)
   const communityId = yield* select(communitiesSelectors.currentCommunityId)
 
   if (certificate) return
@@ -17,12 +17,9 @@ export function* updateCertificateSaga(action: PayloadAction<SendCertificatesRes
   if (!csr?.userCsr) return
 
   const parsedCsr = yield* call(loadCSR, csr?.userCsr)
-  const pubKey = getReqFieldValue(parsedCsr, CertFieldsTypes.dmPublicKey)
 
   const cert = action.payload.certificates.find(cert => {
-    const parsedCert = parseCertificate(cert)
-    const certPubKey = getCertFieldValue(parsedCert, CertFieldsTypes.dmPublicKey)
-    if (certPubKey === pubKey) return cert
+    if (pubKeyMatch(cert, parsedCsr)) return cert
   })
 
   if (cert) {